+++ tests/testthat/test-bind-shadow.R
@@ -49,5 +49,5 @@ test_that(
   "bind_shadow with only_miss = TRUE returns columns with extra suffix _NA",{
   expect_equal(names(bind_shadow(airquality, only_miss = TRUE)),
                c(names(airquality),
-                 paste0(which_var_na(airquality),"_NA")))
+                 paste0(miss_var_which(airquality),"_NA")))
 })