+++ NEWS.md
@@ -1,10 +1,14 @@
+# naniar 0.2.1.9600 (2018/05/02) 
+
+- `gg_miss_which` is rotated 90 degrees to make it easier to read the variables
+- added `miss_var_which`, a function that lists the variables missing
+
 # naniar 0.2.1.9500 (2018/05/01) 
 
 # New features
 
 - `impute_below` and `shadow_shift` gain arguments `prop_below` and `jitter` 
 to control the degree of shift, and also the extent of jitter.
-- `gg_miss_which` is rotated 90 degrees to make it easier to read the variables
 
 # naniar 0.2.1.9400 (2018/04/30)